[US13] Create endpoints for Google+ login and forgot password (#161)
* debug US-13

* [US13] Create endpoints for Google+ login and forgot password

* [US13] Create endpoints for Google+ login and forgot password

+++ social_network_api/src/main/java/com/danit/socialnetwork/rest/PasswordRestController.java
@@ -1,6 +1,9 @@
 package com.danit.socialnetwork.rest;
 
 import com.danit.socialnetwork.controller.PasswordChanger;
+import com.danit.socialnetwork.dto.ChangePasswordRequest;
+import com.danit.socialnetwork.dto.CodeCheckRequest;
+import com.danit.socialnetwork.dto.NewPasswordRequest;
 import com.danit.socialnetwork.model.DbUser;
 import com.danit.socialnetwork.model.PasswordChangeRequests;
 import com.danit.socialnetwork.repository.UserRepository;
@@ -11,15 +14,11 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.GetMapping;
 
 
-import javax.servlet.http.HttpServletRequest;
-import java.util.Map;
 import java.util.Optional;
 
 @Log4j2
@@ -30,20 +29,18 @@ public class PasswordRestController {
   private final PasswordChangerService passwordChangerService;
   private final UserRepository userRepo;
 
-  //    @RequiredArgsConstructor
   @Data
   private static class UserEmail {
     private final String email;
   }
 
-  @PostMapping("/changepassword")
-  public ResponseEntity<?> changePass(@RequestBody Map<String, String> loginData,
-                                      HttpServletRequest request) {
-    String userEmail = loginData.get("email");
+  @PostMapping("/api/changepassword")
+  public ResponseEntity<?> changePass(@RequestBody ChangePasswordRequest changePasswordRequest) {
+    String userEmail = changePasswordRequest.getEmail();
     Optional<DbUser> maybeUser = userRepo.findDbUserByEmail(userEmail);
 
     if (maybeUser.isPresent()) {
-      String secretUrl = passChanger.change(request, userEmail);
+      String secretUrl = passChanger.change(userEmail);
       log.info(passwordChangerService.saveRequest(userEmail, secretUrl));
       return ResponseEntity.ok("Instructions sent on, " + userEmail);
     } else {
@@ -52,27 +49,30 @@ public class PasswordRestController {
     }
   }
 
-  @GetMapping("/changepassword{uuid}")
-  public ResponseEntity<?> res(@PathVariable("uuid") String uuid) {
-
-    Optional<PasswordChangeRequests> maybeRequest = passwordChangerService.getEmailByUuid(uuid);
+  @PostMapping("/api/codecheck")
+  public ResponseEntity<?> codeCheck(@RequestBody CodeCheckRequest codeCheckRequest) {
+    String userEmail = codeCheckRequest.getEmail();
+    String secretCode = codeCheckRequest.getCode();
+    Optional<PasswordChangeRequests> maybeRequest = passwordChangerService.getEmailByUuid(secretCode);
 
     if (maybeRequest.isPresent()) {
       UserEmail email = new UserEmail(maybeRequest.get().getEmail());
-      log.info("Change password page call from e-mail " + email.getEmail());
-      passwordChangerService.deleteRequestByEmail(maybeRequest.get().getEmail());
-      return new ResponseEntity<>(email, HttpStatus.OK);
-    } else {
-      log.info("Invalid Change password page link from e-mail");
-      return new ResponseEntity<>("Invalid Change password page link from e-mail",
-          HttpStatus.BAD_REQUEST);
+      if (email.getEmail().equals(userEmail)) {
+        log.info("Change password page call from e-mail " + email.getEmail());
+        passwordChangerService.deleteRequestByEmail(maybeRequest.get().getEmail());
+        return new ResponseEntity<>("code accessed", HttpStatus.OK);
+      }
     }
+    log.info("Invalid code");
+    return new ResponseEntity<>("Invalid code",
+        HttpStatus.BAD_REQUEST);
   }
 
-  @PostMapping("/newpassword")
-  public ResponseEntity<?> authenticateUser(@RequestBody Map<String, String> loginData) {
-    String userEmail = loginData.get("email");
-    String password = loginData.get("password");
+
+  @PostMapping("/api/newpassword")
+  public ResponseEntity<?> authenticateUser(@RequestBody NewPasswordRequest newPasswordRequest) {
+    String userEmail = newPasswordRequest.getEmail();
+    String password = newPasswordRequest.getPassword();
 
     Optional<DbUser> maybeUser = userRepo.findDbUserByEmail(userEmail);
